Tests and automatic type generation (#134)
This PR does two big things:

1) Automatic generation of types. As part of this, additional types
information has been moved across to some of the source `.ts` files
(mainly around http queries).

2) Adds a wide range of tests, in `tests/`.

There's still a load of clean-up to be done, which is what I will be
working on next. In the meantime, probably best to ignore the contents
of `dist/`, which we'll probably stop committing in future.