Small fixes and dependency updates (#159)
* #154 — asks to export `escapeLiteral` and `escapeIdentifier`: this
does so.

* #112 — points out an issue with Deno/JSR, which is that we pin to a
version of `@types/pg` that does not exist. Oops! This fixes that too.

* The `postversion` script now points out the need to publish to JSR.

* All packages (except `pg`, which is pinned to 8.8.0) are updated,
fixing some security warnings. Prisma (or at least its type definition)
appears to have changed in 6.6.0, so the relevant tests are updated and
simplified to match. Since we now rely on Prisma to import
`@neondatabase/serverless` for itself, I have added an `override` field
in `package.json` to ensure it uses the local package, not whatever
version is currently on npm.